feat: Refactor Kafka cron producer code structure
