+++ projects/kafka-cron/producer/pkg/models/jobs.go
